Moved composer-locator utilities to a `Factory\` sub-namespace
